# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Custom Analytic - WW2 API",
        "description": "API Documentation for Custom Analytic - WW2",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/feedbacks": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Add feedback",
                "operationId": "d32093aef824dfefbb986bc67cf96a27",
                "responses": {
                    "200": {
                        "description": "New feedback ... has been created"
                    },
                    "422": {
                        "description": "{validation_msg}"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/feedbacks/limit/{limit}/order/{order}/{id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Show all feedbacks per stories with pagination, ordering, and search",
                "operationId": "71797070b0659a258a79a230544a470b",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of feedback per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order by field created at",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the stories",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "10e62495-d508-b32f-2e8e-66fa84f43fe8"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "feedback found"
                    },
                    "404": {
                        "description": "feedback not found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/feedbacks/stats/{id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Show stories feedback stats by id",
                "operationId": "8d1f6a4875a1e4611c446dca88365f03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the stories",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "10e62495-d508-b32f-2e8e-66fa84f43fe8"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "feedback stats found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}